+++ skbio/diversity/beta/tests/test_unifrac.py
@@ -24,9 +24,6 @@ from skbio.diversity.beta._unifrac import (_unweighted_unifrac,
 class UnifracTests(TestCase):
 
     def setUp(self):
-        self.unweighted_unifrac = unweighted_unifrac
-        self.weighted_unifrac = weighted_unifrac
-
         self.b1 = np.array(
             [[1, 3, 0, 1, 0],
              [0, 2, 0, 4, 4],
@@ -60,9 +57,9 @@ class UnifracTests(TestCase):
 
         for i in range(len(self.b1)):
             for j in range(len(self.b1)):
-                actual = self.unweighted_unifrac(
+                actual = unweighted_unifrac(
                     self.b1[i], self.b1[j], self.oids1, self.t1)
-                expected = self.unweighted_unifrac(
+                expected = unweighted_unifrac(
                     shuffled_b1[i], shuffled_b1[j], shuffled_ids, self.t1)
                 self.assertAlmostEqual(actual, expected)
 
@@ -77,9 +74,9 @@ class UnifracTests(TestCase):
 
         for i in range(len(self.b1)):
             for j in range(len(self.b1)):
-                actual = self.weighted_unifrac(
+                actual = weighted_unifrac(
                     self.b1[i], self.b1[j], self.oids1, self.t1)
-                expected = self.weighted_unifrac(
+                expected = weighted_unifrac(
                     shuffled_b1[i], shuffled_b1[j], shuffled_ids, self.t1)
                 self.assertAlmostEqual(actual, expected)
 
@@ -87,9 +84,9 @@ class UnifracTests(TestCase):
         # UniFrac values are the same despite unobserved tips in the tree
         for i in range(len(self.b1)):
             for j in range(len(self.b1)):
-                actual = self.unweighted_unifrac(
+                actual = unweighted_unifrac(
                     self.b1[i], self.b1[j], self.oids1, self.t1_w_extra_tips)
-                expected = self.unweighted_unifrac(
+                expected = unweighted_unifrac(
                     self.b1[i], self.b1[j], self.oids1, self.t1)
                 self.assertAlmostEqual(actual, expected)
 
@@ -97,24 +94,24 @@ class UnifracTests(TestCase):
         # UniFrac values are the same despite unobserved tips in the tree
         for i in range(len(self.b1)):
             for j in range(len(self.b1)):
-                actual = self.weighted_unifrac(
+                actual = weighted_unifrac(
                     self.b1[i], self.b1[j], self.oids1, self.t1_w_extra_tips)
-                expected = self.weighted_unifrac(
+                expected = weighted_unifrac(
                     self.b1[i], self.b1[j], self.oids1, self.t1)
                 self.assertAlmostEqual(actual, expected)
 
     def test_unweighted_minimal_trees(self):
         # two tips
         tree = TreeNode.read(StringIO(u'(OTU1:0.25, OTU2:0.25)root;'))
-        actual = self.unweighted_unifrac([1, 0], [0, 0], ['OTU1', 'OTU2'],
-                                         tree)
+        actual = unweighted_unifrac([1, 0], [0, 0], ['OTU1', 'OTU2'],
+                                    tree)
         expected = 1.0
         self.assertEqual(actual, expected)
 
     def test_weighted_minimal_trees(self):
         # two tips
         tree = TreeNode.read(StringIO(u'(OTU1:0.25, OTU2:0.25)root;'))
-        actual = self.weighted_unifrac([1, 0], [0, 0], ['OTU1', 'OTU2'], tree)
+        actual = weighted_unifrac([1, 0], [0, 0], ['OTU1', 'OTU2'], tree)
         expected = 0.25
         self.assertEqual(actual, expected)
 
@@ -122,8 +119,8 @@ class UnifracTests(TestCase):
         # expected values computed with QIIME 1.9.1 and by hand
         # root node not observed, but branch between (OTU1, OTU2) and root
         # is considered shared
-        actual = self.unweighted_unifrac([1, 1, 0, 0], [1, 0, 0, 0],
-                                         self.oids2, self.t2)
+        actual = unweighted_unifrac([1, 1, 0, 0], [1, 0, 0, 0],
+                                    self.oids2, self.t2)
         # for clarity of what I'm testing, compute expected as it would
         # based on the branch lengths. the values that compose shared was
         # a point of confusion for me here, so leaving these in for
@@ -133,8 +130,8 @@ class UnifracTests(TestCase):
 
         # root node not observed, but branch between (OTU3, OTU4) and root
         # is considered shared
-        actual = self.unweighted_unifrac([0, 0, 1, 1], [0, 0, 1, 0],
-                                         self.oids2, self.t2)
+        actual = unweighted_unifrac([0, 0, 1, 1], [0, 0, 1, 0],
+                                    self.oids2, self.t2)
         # for clarity of what I'm testing, compute expected as it would
         # based on the branch lengths. the values that compose shared was
         # a point of confusion for me here, so leaving these in for
@@ -146,15 +143,15 @@ class UnifracTests(TestCase):
         # expected values computed by hand, these disagree with QIIME 1.9.1
         # root node not observed, but branch between (OTU1, OTU2) and root
         # is considered shared
-        actual = self.weighted_unifrac([1, 0, 0, 0], [1, 1, 0, 0],
-                                       self.oids2, self.t2)
+        actual = weighted_unifrac([1, 0, 0, 0], [1, 1, 0, 0],
+                                  self.oids2, self.t2)
         expected = 0.15
         self.assertAlmostEqual(actual, expected)
 
         # root node not observed, but branch between (OTU3, OTU4) and root
         # is considered shared
-        actual = self.weighted_unifrac([0, 0, 1, 1], [0, 0, 1, 0],
-                                       self.oids2, self.t2)
+        actual = weighted_unifrac([0, 0, 1, 1], [0, 0, 1, 0],
+                                  self.oids2, self.t2)
         expected = 0.6
         self.assertAlmostEqual(actual, expected)
 
@@ -162,21 +159,21 @@ class UnifracTests(TestCase):
         # expected values computed by hand, these disagree with QIIME 1.9.1
         # root node not observed, but branch between (OTU1, OTU2) and root
         # is considered shared
-        actual = self.weighted_unifrac([1, 0, 0, 0], [1, 1, 0, 0],
-                                       self.oids2, self.t2, normalized=True)
+        actual = weighted_unifrac([1, 0, 0, 0], [1, 1, 0, 0],
+                                  self.oids2, self.t2, normalized=True)
         expected = 0.1764705882
         self.assertAlmostEqual(actual, expected)
 
         # root node not observed, but branch between (OTU3, OTU4) and root
         # is considered shared
-        actual = self.weighted_unifrac([0, 0, 1, 1], [0, 0, 1, 0],
-                                       self.oids2, self.t2, normalized=True)
+        actual = weighted_unifrac([0, 0, 1, 1], [0, 0, 1, 0],
+                                  self.oids2, self.t2, normalized=True)
         expected = 0.1818181818
         self.assertAlmostEqual(actual, expected)
 
     def test_unweighted_unifrac_identity(self):
         for i in range(len(self.b1)):
-            actual = self.unweighted_unifrac(
+            actual = unweighted_unifrac(
                 self.b1[i], self.b1[i], self.oids1, self.t1)
             expected = 0.0
             self.assertAlmostEqual(actual, expected)
@@ -184,9 +181,9 @@ class UnifracTests(TestCase):
     def test_unweighted_unifrac_symmetry(self):
         for i in range(len(self.b1)):
             for j in range(len(self.b1)):
-                actual = self.unweighted_unifrac(
+                actual = unweighted_unifrac(
                     self.b1[i], self.b1[j], self.oids1, self.t1)
-                expected = self.unweighted_unifrac(
+                expected = unweighted_unifrac(
                     self.b1[j], self.b1[i], self.oids1, self.t1)
                 self.assertAlmostEqual(actual, expected)
 
@@ -202,9 +199,9 @@ class UnifracTests(TestCase):
         u_counts = [1, 2, 3]
         v_counts = [1, 1, 1]
         otu_ids = ['OTU1', 'OTU2', 'OTU3']
-        self.assertRaises(DuplicateNodeError, self.unweighted_unifrac,
+        self.assertRaises(DuplicateNodeError, unweighted_unifrac,
                           u_counts, v_counts, otu_ids, t)
-        self.assertRaises(DuplicateNodeError, self.weighted_unifrac, u_counts,
+        self.assertRaises(DuplicateNodeError, weighted_unifrac, u_counts,
                           v_counts, otu_ids, t)
 
         # unrooted tree as input
@@ -213,9 +210,9 @@ class UnifracTests(TestCase):
         u_counts = [1, 2, 3]
         v_counts = [1, 1, 1]
         otu_ids = ['OTU1', 'OTU2', 'OTU3']
-        self.assertRaises(ValueError, self.unweighted_unifrac, u_counts,
+        self.assertRaises(ValueError, unweighted_unifrac, u_counts,
                           v_counts, otu_ids, t)
-        self.assertRaises(ValueError, self.weighted_unifrac, u_counts,
+        self.assertRaises(ValueError, weighted_unifrac, u_counts,
                           v_counts, otu_ids, t)
 
         # otu_ids has duplicated ids
@@ -225,9 +222,9 @@ class UnifracTests(TestCase):
         u_counts = [1, 2, 3]
         v_counts = [1, 1, 1]
         otu_ids = ['OTU1', 'OTU2', 'OTU2']
-        self.assertRaises(ValueError, self.unweighted_unifrac, u_counts,
+        self.assertRaises(ValueError, unweighted_unifrac, u_counts,
                           v_counts, otu_ids, t)
-        self.assertRaises(ValueError, self.weighted_unifrac, u_counts,
+        self.assertRaises(ValueError, weighted_unifrac, u_counts,
                           v_counts, otu_ids, t)
 
         # len of vectors not equal
@@ -237,23 +234,23 @@ class UnifracTests(TestCase):
         u_counts = [1, 2]
         v_counts = [1, 1, 1]
         otu_ids = ['OTU1', 'OTU2', 'OTU3']
-        self.assertRaises(ValueError, self.unweighted_unifrac, u_counts,
+        self.assertRaises(ValueError, unweighted_unifrac, u_counts,
                           v_counts, otu_ids, t)
-        self.assertRaises(ValueError, self.weighted_unifrac, u_counts,
+        self.assertRaises(ValueError, weighted_unifrac, u_counts,
                           v_counts, otu_ids, t)
         u_counts = [1, 2, 3]
         v_counts = [1, 1]
         otu_ids = ['OTU1', 'OTU2', 'OTU3']
-        self.assertRaises(ValueError, self.unweighted_unifrac, u_counts,
+        self.assertRaises(ValueError, unweighted_unifrac, u_counts,
                           v_counts, otu_ids, t)
-        self.assertRaises(ValueError, self.weighted_unifrac, u_counts,
+        self.assertRaises(ValueError, weighted_unifrac, u_counts,
                           v_counts, otu_ids, t)
         u_counts = [1, 2, 3]
         v_counts = [1, 1, 1]
         otu_ids = ['OTU1', 'OTU2']
-        self.assertRaises(ValueError, self.unweighted_unifrac, u_counts,
+        self.assertRaises(ValueError, unweighted_unifrac, u_counts,
                           v_counts, otu_ids, t)
-        self.assertRaises(ValueError, self.weighted_unifrac, u_counts,
+        self.assertRaises(ValueError, weighted_unifrac, u_counts,
                           v_counts, otu_ids, t)
 
         # negative counts
@@ -263,16 +260,16 @@ class UnifracTests(TestCase):
         u_counts = [1, 2, -3]
         v_counts = [1, 1, 1]
         otu_ids = ['OTU1', 'OTU2', 'OTU3']
-        self.assertRaises(ValueError, self.unweighted_unifrac, u_counts,
+        self.assertRaises(ValueError, unweighted_unifrac, u_counts,
                           v_counts, otu_ids, t)
-        self.assertRaises(ValueError, self.weighted_unifrac, u_counts,
+        self.assertRaises(ValueError, weighted_unifrac, u_counts,
                           v_counts, otu_ids, t)
         u_counts = [1, 2, 3]
         v_counts = [1, 1, -1]
         otu_ids = ['OTU1', 'OTU2', 'OTU3']
-        self.assertRaises(ValueError, self.unweighted_unifrac, u_counts,
+        self.assertRaises(ValueError, unweighted_unifrac, u_counts,
                           v_counts, otu_ids, t)
-        self.assertRaises(ValueError, self.weighted_unifrac, u_counts,
+        self.assertRaises(ValueError, weighted_unifrac, u_counts,
                           v_counts, otu_ids, t)
 
         # tree with no branch lengths
@@ -281,9 +278,9 @@ class UnifracTests(TestCase):
         u_counts = [1, 2, 3]
         v_counts = [1, 1, 1]
         otu_ids = ['OTU1', 'OTU2', 'OTU3']
-        self.assertRaises(ValueError, self.unweighted_unifrac, u_counts,
+        self.assertRaises(ValueError, unweighted_unifrac, u_counts,
                           v_counts, otu_ids, t)
-        self.assertRaises(ValueError, self.weighted_unifrac, u_counts,
+        self.assertRaises(ValueError, weighted_unifrac, u_counts,
                           v_counts, otu_ids, t)
 
         # tree missing some branch lengths
@@ -293,9 +290,9 @@ class UnifracTests(TestCase):
         u_counts = [1, 2, 3]
         v_counts = [1, 1, 1]
         otu_ids = ['OTU1', 'OTU2', 'OTU3']
-        self.assertRaises(ValueError, self.unweighted_unifrac, u_counts,
+        self.assertRaises(ValueError, unweighted_unifrac, u_counts,
                           v_counts, otu_ids, t)
-        self.assertRaises(ValueError, self.weighted_unifrac, u_counts,
+        self.assertRaises(ValueError, weighted_unifrac, u_counts,
                           v_counts, otu_ids, t)
 
         # otu_ids not present in tree
@@ -305,32 +302,32 @@ class UnifracTests(TestCase):
         u_counts = [1, 2, 3]
         v_counts = [1, 1, 1]
         otu_ids = ['OTU1', 'OTU2', 'OTU42']
-        self.assertRaises(MissingNodeError, self.unweighted_unifrac, u_counts,
+        self.assertRaises(MissingNodeError, unweighted_unifrac, u_counts,
                           v_counts, otu_ids, t)
-        self.assertRaises(MissingNodeError, self.weighted_unifrac, u_counts,
+        self.assertRaises(MissingNodeError, weighted_unifrac, u_counts,
                           v_counts, otu_ids, t)
 
     def test_unweighted_unifrac_non_overlapping(self):
         # these communities only share the root node
-        actual = self.unweighted_unifrac(
+        actual = unweighted_unifrac(
             self.b1[4], self.b1[5], self.oids1, self.t1)
         expected = 1.0
         self.assertAlmostEqual(actual, expected)
-        actual = self.unweighted_unifrac(
+        actual = unweighted_unifrac(
             [1, 1, 1, 0, 0], [0, 0, 0, 1, 1], self.oids1, self.t1)
         expected = 1.0
         self.assertAlmostEqual(actual, expected)
 
     def test_unweighted_unifrac_zero_counts(self):
-        actual = self.unweighted_unifrac(
+        actual = unweighted_unifrac(
             [1, 1, 1, 0, 0], [0, 0, 0, 0, 0], self.oids1, self.t1)
         expected = 1.0
         self.assertAlmostEqual(actual, expected)
-        actual = self.unweighted_unifrac(
+        actual = unweighted_unifrac(
             [0, 0, 0, 0, 0], [0, 0, 0, 0, 0], self.oids1, self.t1)
         expected = 0.0
         self.assertAlmostEqual(actual, expected)
-        actual = self.unweighted_unifrac(
+        actual = unweighted_unifrac(
             [], [], [], self.t1)
         expected = 0.0
         self.assertAlmostEqual(actual, expected)
@@ -340,74 +337,74 @@ class UnifracTests(TestCase):
         # is a completely different implementation skbio's initial
         # unweighted unifrac implementation
         # sample A versus all
-        actual = self.unweighted_unifrac(
+        actual = unweighted_unifrac(
             self.b1[0], self.b1[1], self.oids1, self.t1)
         expected = 0.238095238095
         self.assertAlmostEqual(actual, expected)
-        actual = self.unweighted_unifrac(
+        actual = unweighted_unifrac(
             self.b1[0], self.b1[2], self.oids1, self.t1)
         expected = 0.52
         self.assertAlmostEqual(actual, expected)
-        actual = self.unweighted_unifrac(
+        actual = unweighted_unifrac(
             self.b1[0], self.b1[3], self.oids1, self.t1)
         expected = 0.52
         self.assertAlmostEqual(actual, expected)
-        actual = self.unweighted_unifrac(
+        actual = unweighted_unifrac(
             self.b1[0], self.b1[4], self.oids1, self.t1)
         expected = 0.545454545455
         self.assertAlmostEqual(actual, expected)
-        actual = self.unweighted_unifrac(
+        actual = unweighted_unifrac(
             self.b1[0], self.b1[5], self.oids1, self.t1)
         expected = 0.619047619048
         self.assertAlmostEqual(actual, expected)
         # sample B versus remaining
-        actual = self.unweighted_unifrac(
+        actual = unweighted_unifrac(
             self.b1[1], self.b1[2], self.oids1, self.t1)
         expected = 0.347826086957
         self.assertAlmostEqual(actual, expected)
-        actual = self.unweighted_unifrac(
+        actual = unweighted_unifrac(
             self.b1[1], self.b1[3], self.oids1, self.t1)
         expected = 0.347826086957
         self.assertAlmostEqual(actual, expected)
-        actual = self.unweighted_unifrac(
+        actual = unweighted_unifrac(
             self.b1[1], self.b1[4], self.oids1, self.t1)
         expected = 0.68
         self.assertAlmostEqual(actual, expected)
-        actual = self.unweighted_unifrac(
+        actual = unweighted_unifrac(
             self.b1[1], self.b1[5], self.oids1, self.t1)
         expected = 0.421052631579
         self.assertAlmostEqual(actual, expected)
         # sample C versus remaining
-        actual = self.unweighted_unifrac(
+        actual = unweighted_unifrac(
             self.b1[2], self.b1[3], self.oids1, self.t1)
         expected = 0.0
         self.assertAlmostEqual(actual, expected)
-        actual = self.unweighted_unifrac(
+        actual = unweighted_unifrac(
             self.b1[2], self.b1[4], self.oids1, self.t1)
         expected = 0.68
         self.assertAlmostEqual(actual, expected)
-        actual = self.unweighted_unifrac(
+        actual = unweighted_unifrac(
             self.b1[2], self.b1[5], self.oids1, self.t1)
         expected = 0.421052631579
         self.assertAlmostEqual(actual, expected)
         # sample D versus remaining
-        actual = self.unweighted_unifrac(
+        actual = unweighted_unifrac(
             self.b1[3], self.b1[4], self.oids1, self.t1)
         expected = 0.68
         self.assertAlmostEqual(actual, expected)
-        actual = self.unweighted_unifrac(
+        actual = unweighted_unifrac(
             self.b1[3], self.b1[5], self.oids1, self.t1)
         expected = 0.421052631579
         self.assertAlmostEqual(actual, expected)
         # sample E versus remaining
-        actual = self.unweighted_unifrac(
+        actual = unweighted_unifrac(
             self.b1[4], self.b1[5], self.oids1, self.t1)
         expected = 1.0
         self.assertAlmostEqual(actual, expected)
 
     def test_weighted_unifrac_identity(self):
         for i in range(len(self.b1)):
-            actual = self.weighted_unifrac(
+            actual = weighted_unifrac(
                 self.b1[i], self.b1[i], self.oids1, self.t1)
             expected = 0.0
             self.assertAlmostEqual(actual, expected)
@@ -415,9 +412,9 @@ class UnifracTests(TestCase):
     def test_weighted_unifrac_symmetry(self):
         for i in range(len(self.b1)):
             for j in range(len(self.b1)):
-                actual = self.weighted_unifrac(
+                actual = weighted_unifrac(
                     self.b1[i], self.b1[j], self.oids1, self.t1)
-                expected = self.weighted_unifrac(
+                expected = weighted_unifrac(
                     self.b1[j], self.b1[i], self.oids1, self.t1)
                 self.assertAlmostEqual(actual, expected)
 
@@ -426,24 +423,24 @@ class UnifracTests(TestCase):
         # is a completely different implementation skbio's initial
         # weighted unifrac implementation
         # these communities only share the root node
-        actual = self.weighted_unifrac(
+        actual = weighted_unifrac(
             self.b1[4], self.b1[5], self.oids1, self.t1)
         expected = 4.0
         self.assertAlmostEqual(actual, expected)
 
     def test_weighted_unifrac_zero_counts(self):
-        actual = self.weighted_unifrac(
+        actual = weighted_unifrac(
             [0, 0, 0, 0, 0], [0, 0, 0, 0, 0], self.oids1, self.t1)
         expected = 0.0
         self.assertAlmostEqual(actual, expected)
         # calculated the following by hand, as QIIME 1.9.1 tells the user
         # that values involving empty vectors will be uninformative, and
         # returns 1.0
-        actual = self.weighted_unifrac(
+        actual = weighted_unifrac(
             [1, 1, 1, 0, 0], [0, 0, 0, 0, 0], self.oids1, self.t1)
         expected = 2.0
         self.assertAlmostEqual(actual, expected)
-        actual = self.weighted_unifrac(
+        actual = weighted_unifrac(
             [], [], [], self.t1)
         expected = 0.0
         self.assertAlmostEqual(actual, expected)
@@ -452,74 +449,74 @@ class UnifracTests(TestCase):
         # expected results derived from QIIME 1.9.1, which
         # is a completely different implementation skbio's initial
         # weighted unifrac implementation
-        actual = self.weighted_unifrac(
+        actual = weighted_unifrac(
             self.b1[0], self.b1[1], self.oids1, self.t1)
         expected = 2.4
         self.assertAlmostEqual(actual, expected)
-        actual = self.weighted_unifrac(
+        actual = weighted_unifrac(
             self.b1[0], self.b1[2], self.oids1, self.t1)
         expected = 1.86666666667
         self.assertAlmostEqual(actual, expected)
-        actual = self.weighted_unifrac(
+        actual = weighted_unifrac(
             self.b1[0], self.b1[3], self.oids1, self.t1)
         expected = 2.53333333333
         self.assertAlmostEqual(actual, expected)
-        actual = self.weighted_unifrac(
+        actual = weighted_unifrac(
             self.b1[0], self.b1[4], self.oids1, self.t1)
         expected = 1.35384615385
         self.assertAlmostEqual(actual, expected)
-        actual = self.weighted_unifrac(
+        actual = weighted_unifrac(
             self.b1[0], self.b1[5], self.oids1, self.t1)
         expected = 3.2
         self.assertAlmostEqual(actual, expected)
         # sample B versus remaining
-        actual = self.weighted_unifrac(
+        actual = weighted_unifrac(
             self.b1[1], self.b1[2], self.oids1, self.t1)
         expected = 2.26666666667
         self.assertAlmostEqual(actual, expected)
-        actual = self.weighted_unifrac(
+        actual = weighted_unifrac(
             self.b1[1], self.b1[3], self.oids1, self.t1)
         expected = 0.933333333333
         self.assertAlmostEqual(actual, expected)
-        actual = self.weighted_unifrac(
+        actual = weighted_unifrac(
             self.b1[1], self.b1[4], self.oids1, self.t1)
         expected = 3.2
         self.assertAlmostEqual(actual, expected)
-        actual = self.weighted_unifrac(
+        actual = weighted_unifrac(
             self.b1[1], self.b1[5], self.oids1, self.t1)
         expected = 0.8375
         self.assertAlmostEqual(actual, expected)
         # sample C versus remaining
-        actual = self.weighted_unifrac(
+        actual = weighted_unifrac(
             self.b1[2], self.b1[3], self.oids1, self.t1)
         expected = 1.33333333333
         self.assertAlmostEqual(actual, expected)
-        actual = self.weighted_unifrac(
+        actual = weighted_unifrac(
             self.b1[2], self.b1[4], self.oids1, self.t1)
         expected = 1.89743589744
         self.assertAlmostEqual(actual, expected)
-        actual = self.weighted_unifrac(
+        actual = weighted_unifrac(
             self.b1[2], self.b1[5], self.oids1, self.t1)
         expected = 2.66666666667
         self.assertAlmostEqual(actual, expected)
         # sample D versus remaining
-        actual = self.weighted_unifrac(
+        actual = weighted_unifrac(
             self.b1[3], self.b1[4], self.oids1, self.t1)
         expected = 2.66666666667
         self.assertAlmostEqual(actual, expected)
-        actual = self.weighted_unifrac(
+        actual = weighted_unifrac(
             self.b1[3], self.b1[5], self.oids1, self.t1)
         expected = 1.33333333333
         self.assertAlmostEqual(actual, expected)
         # sample E versus remaining
-        actual = self.weighted_unifrac(
+        actual = weighted_unifrac(
             self.b1[4], self.b1[5], self.oids1, self.t1)
         expected = 4.0
         self.assertAlmostEqual(actual, expected)
 
     def test_weighted_unifrac_identity_normalized(self):
         for i in range(len(self.b1)):
-            actual = self.weighted_unifrac(
+            actual = weighted_unifrac(
                 self.b1[i], self.b1[i], self.oids1, self.t1, normalized=True)
             expected = 0.0
             self.assertAlmostEqual(actual, expected)
@@ -527,21 +524,21 @@ class UnifracTests(TestCase):
     def test_weighted_unifrac_symmetry_normalized(self):
         for i in range(len(self.b1)):
             for j in range(len(self.b1)):
-                actual = self.weighted_unifrac(
+                actual = weighted_unifrac(
                     self.b1[i], self.b1[j], self.oids1, self.t1,
                     normalized=True)
-                expected = self.weighted_unifrac(
+                expected = weighted_unifrac(
                     self.b1[j], self.b1[i], self.oids1, self.t1,
                     normalized=True)
                 self.assertAlmostEqual(actual, expected)
 
     def test_weighted_unifrac_non_overlapping_normalized(self):
         # these communities only share the root node
-        actual = self.weighted_unifrac(
+        actual = weighted_unifrac(
             self.b1[4], self.b1[5], self.oids1, self.t1, normalized=True)
         expected = 1.0
         self.assertAlmostEqual(actual, expected)
-        actual = self.weighted_unifrac(
+        actual = weighted_unifrac(
             [1, 1, 1, 0, 0], [0, 0, 0, 1, 1], self.oids1, self.t1,
             normalized=True)
         expected = 1.0
@@ -551,17 +548,17 @@ class UnifracTests(TestCase):
         # expected results derived from QIIME 1.9.1, which
         # is a completely different implementation skbio's initial
         # weighted unifrac implementation
-        actual = self.weighted_unifrac(
+        actual = weighted_unifrac(
             [0, 0, 0, 0, 0], [0, 0, 0, 0, 0], self.oids1, self.t1,
             normalized=True)
         expected = 0.0
         self.assertAlmostEqual(actual, expected)
-        actual = self.weighted_unifrac(
+        actual = weighted_unifrac(
             [1, 1, 1, 0, 0], [0, 0, 0, 0, 0], self.oids1, self.t1,
             normalized=True)
         expected = 1.0
         self.assertAlmostEqual(actual, expected)
-        actual = self.weighted_unifrac(
+        actual = weighted_unifrac(
             [], [], [], self.t1, normalized=True)
         expected = 0.0
         self.assertAlmostEqual(actual, expected)
@@ -570,67 +567,67 @@ class UnifracTests(TestCase):
         # expected results derived from QIIME 1.9.1, which
         # is a completely different implementation skbio's initial
         # weighted unifrac implementation
-        actual = self.weighted_unifrac(
+        actual = weighted_unifrac(
             self.b1[0], self.b1[1], self.oids1, self.t1, normalized=True)
         expected = 0.6
         self.assertAlmostEqual(actual, expected)
-        actual = self.weighted_unifrac(
+        actual = weighted_unifrac(
             self.b1[0], self.b1[2], self.oids1, self.t1, normalized=True)
         expected = 0.466666666667
         self.assertAlmostEqual(actual, expected)
-        actual = self.weighted_unifrac(
+        actual = weighted_unifrac(
             self.b1[0], self.b1[3], self.oids1, self.t1, normalized=True)
         expected = 0.633333333333
         self.assertAlmostEqual(actual, expected)
-        actual = self.weighted_unifrac(
+        actual = weighted_unifrac(
             self.b1[0], self.b1[4], self.oids1, self.t1, normalized=True)
         expected = 0.338461538462
         self.assertAlmostEqual(actual, expected)
-        actual = self.weighted_unifrac(
+        actual = weighted_unifrac(
             self.b1[0], self.b1[5], self.oids1, self.t1, normalized=True)
         expected = 0.8
         self.assertAlmostEqual(actual, expected)
         # sample B versus remaining
-        actual = self.weighted_unifrac(
+        actual = weighted_unifrac(
             self.b1[1], self.b1[2], self.oids1, self.t1, normalized=True)
         expected = 0.566666666667
         self.assertAlmostEqual(actual, expected)
-        actual = self.weighted_unifrac(
+        actual = weighted_unifrac(
             self.b1[1], self.b1[3], self.oids1, self.t1, normalized=True)
         expected = 0.233333333333
         self.assertAlmostEqual(actual, expected)
-        actual = self.weighted_unifrac(
+        actual = weighted_unifrac(
             self.b1[1], self.b1[4], self.oids1, self.t1, normalized=True)
         expected = 0.8
         self.assertAlmostEqual(actual, expected)
-        actual = self.weighted_unifrac(
+        actual = weighted_unifrac(
             self.b1[1], self.b1[5], self.oids1, self.t1, normalized=True)
         expected = 0.209375
         self.assertAlmostEqual(actual, expected)
         # sample C versus remaining
-        actual = self.weighted_unifrac(
+        actual = weighted_unifrac(
             self.b1[2], self.b1[3], self.oids1, self.t1, normalized=True)
         expected = 0.333333333333
         self.assertAlmostEqual(actual, expected)
-        actual = self.weighted_unifrac(
+        actual = weighted_unifrac(
             self.b1[2], self.b1[4], self.oids1, self.t1, normalized=True)
         expected = 0.474358974359
         self.assertAlmostEqual(actual, expected)
-        actual = self.weighted_unifrac(
+        actual = weighted_unifrac(
             self.b1[2], self.b1[5], self.oids1, self.t1, normalized=True)
         expected = 0.666666666667
         self.assertAlmostEqual(actual, expected)
         # sample D versus remaining
-        actual = self.weighted_unifrac(
+        actual = weighted_unifrac(
             self.b1[3], self.b1[4], self.oids1, self.t1, normalized=True)
         expected = 0.666666666667
         self.assertAlmostEqual(actual, expected)
-        actual = self.weighted_unifrac(
+        actual = weighted_unifrac(
             self.b1[3], self.b1[5], self.oids1, self.t1, normalized=True)
         expected = 0.333333333333
         self.assertAlmostEqual(actual, expected)
         # sample E versus remaining
-        actual = self.weighted_unifrac(
+        actual = weighted_unifrac(
             self.b1[4], self.b1[5], self.oids1, self.t1, normalized=True)
         expected = 1.0
         self.assertAlmostEqual(actual, expected)